Add conditional compilation and linking
In order to get this to build on macOS (and LLVM compilers version 9 or
above) we need to conditionally link the c++fs library instead of always
including it. Before version 9 the filesystem library was not a standard
part, however it is starting at version 9 and above.

Additionally, for non-Alsa systems we need to forward declare the
structures that are in use to avoid build errors.