Remove Old Clang Gating & Unwind Forwarding
Since we're pushing for Clang 10 we can remove the conditional linking
all together as it will not be needed at all. Additionally, as mentioned
in PR feedback, the forwarding of ALSA structures is the wrong way to
handle this conditional approach since it was a hack that should be
fixed in a different way.