Update Travis To Clang 10
We're currently supporting on Clang 10+ for building this project, so we
should clean up our CI jobs to make sure they are using Clang 10 to
catch relevant errors and to match local build environments.

Additionally, it seems like LLVM recently moved all new releases to
being published in Github so we have to change our install scripts to
point to them to download the Clang 10 sources.