Upgrade Travis To Newer Ubuntu Distribution & GCC 10 Upgrade (#177)
* Upgrade Travis to Focal

Focal Fossa (Ubuntu 20.04) is the latest LTS that was released recently
and will be supported through April 2025. It also seems that Travis is
supporting this, but haven't updated their docs accordingly. This will
bring forward our CI environment to easily give access to new mainline
packages for gcc-10 and more.

* move to sourceline: configuration

* Use latest officially supported OS

* Upgrade to GCC 10

* Allow clang/libstdc++ to fail as it currently doesnt work with floating point atomics

Co-authored-by: topisani <topisani@hamsterpoison.com>